Replace link to unreachable remergr.io with another project (#10)
I stumbled into this repo while I was doing some research before working on https://github.com/tibdex/autorebase. I clicked on the link to remergr.io but it has become unreachable. I'm proposing replacing the link to a one pointing to Autorebase. This can be seen as a shameless but I believe Autorebase is close to the original goals of this project.

Feel free to merge or close ;)